+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Undefined, Null functions to get these constant values for the isolate
 - Support for calling a method on an object.
 - Support for calling `IsExecutionTerminating` on isolate to check if execution is still terminating.
+- Support for setting and getting internal fields for template object instances
 
 ### Changed
 - Removed error return value from NewIsolate which never fails